Rename `surfaces` module to `surface`
This better matches the names of its sibling modules.
